refactor(wizard): consolidate service pages into WizardServicePages.swift; remove originals (no public API changes)
